# kvs: remove client disconnect bottleneck