@chu11 2021-05-19
> Unlike streaming RPCs that have to be removed to avoid leaking state, these requests will eventually be handled. Silly question: what if we just let that happen?

In other words, just don't bother calling this path at all during a disconnect?  Hmmm.

I suppose my one concern would be the corner case when the `disconnect` is legit, like if a node goes down under a system instance case.  Don't we have to handle that?