@garlick 2021-05-19
> In other words, just don't bother calling this path at all during a disconnect? Hmmm.

Yup :-)  It would probably require a comment in the code e.g.
```c
/* Don't bother finding local cache entries with pending requests from
 * disconnecting peers because it's more trouble than it's worth to cancel them.
 * Let the responses be sent to the abyss.
 */
```

I guess I'm proposing it as a possibility and wanted to get your opinion.

> I suppose my one concern would be the corner case when the disconnect is legit, like if a node goes down under a system instance case. Don't we have to handle that?

This doesn't currently happen, but besides that, I think the fact that these requests are short lived and gone once answered still means that we can send those responses without ill effects.  The uuid that sent the requests won't be reused so they will never go anywhere to cause confusion.