@chu11 2021-05-19
>  This doesn't currently happen, but besides that, I think the fact that these requests are short lived and gone once answered still means that we can send those responses without ill effects. The uuid that sent the requests won't be reused so they will never go anywhere to cause confusion.

Ahhh, got it got it.  I was a bit confused b/c I was thinking the wait queue also stored messages destined for the content-store.  But that's not the case.  Ever since `kvs-watch` was created, the only wait queue messages are general KVS lookups, e.g. short lived as you said.

So I think that'll work.  We can just drop the disconnect cleanup and perhaps some code will become dead code as a result.