@chu11 2021-05-20
Just re-pushed.  So reviewing code and I've been going back and forth on this.

I dislike removing the call to `cache_wait_destroy_msg()`, as it sort of eliminates some abstraction.  The only reason we know we don't have to call it is b/c we happen to know the only messages stored in the cache are lookup RPCs that happen to be short lived.  But the `cache` and `waitqueue` have no idea what types of messages are being stored in it.  So the removal of the call is sort of dependent on knowledge several layers higher than the call.  If the KVS were ever to change, then we would need to re-add `cache_wait_destroy_msg()`.

So I ended up removing the call but added this comment:

```
+    /* Messages can be stored in wait objects via
+     * wait_create_msg_handler(), and subsequently added into the
+     * cache via cache_entry_wait_notdirty() or
+     * cache_entry_wait_valid().
+     *
+     * At the moment, the only messages stored in the cache are KVS
+     * lookups.  Since these are short lived messages, we won't bother
+     * to clean them up in the cache.  When the KVS attempts to
+     * respond to a message from a disconnected client, let the
+     * message simply go to the abyss.  If a longer lived message is
+     * ever added into the cache, the call to cache_wait_destroy_msg()
+     * should be re-added.
+     */
```

Which strongly indicates that the call needs to be re-added if the messages stored change.

I then added another commit to remove `cache_wait_destroy_msg()`, and `wait_destroy_msg()` as the code is no longer used.  Hopefully the above comment would make clear to a future individual to dig up the code that once housed `cache_wait_destroy_msg()`.

In the event that `cache_wait_destroy_msg()` ever has to be brought back, I left my original commit in that started this PR.  Hope that isn't weird.  As I basically remove a commit that I added in the same PR.