@chu11 2021-05-20
Oh yeah, one other idea I pondered yesterday.  The `wait_destroy_msg()` function only works against waits that have messages attached to them.  Not all waits have messages attached to them.  If I were to do some extra counting inside wait queue objects and add a `wait_queue_length_with_messages()` (or something like that), then the iteration in `cache_wait_destroy_msg()` would be even lower than it is now.

Now that I slept on it, I'm actually leaning more towards this implementation now.  If we can get the iteration extremely low, perhaps its best to keep the disconnect cleanup in there.