@garlick 2021-05-20
I was going to say that you could add a check that `flux_msg_is_streaming()` returns false at the point where messages are added to the cache entries to enforce the new invariant.  But I'm OK if you want to go the other way.

I think my main concern is that we don't add/maintain complexity disproportionate with its benefit.