@chu11 2021-05-20
> I was going to say that you could add a check that flux_msg_is_streaming() returns false at the point where messages are added to the cache entries to enforce the new invariant.

One of the things that made me waffle was the `kvssync` code (i.e. support for `flux kvs wait version`).  Those messages were in that in between zone where the message may live in the KVS's memory for a long time if a client disconnects.  I could chose to let responses goto the abyss whenever the wait completes, or should I continue to clean them up?

If I removed all disconnect cleanup, I might have leaned one direction on the code.  But felt that the kvssync disconnect cleanup should probably stay.