@chu11 2021-05-21
re-pushed.  Per conversation above, I've re-done the patch a bit.  I've added a "msg count" set of logic into the waitqueue.  B/c `wait_destroy_msg()` only operates on messages, we can use this even further to limit iteration in `cache_wait_destroy_msg()`.

The original patch is not so different, but now we use `wait_queue_msgs_count()`  instead of `wait_queue_length()`.